Fix handling of both client/server generators
