+++ codegen/src/main/scala/akka/grpc/gen/Main.scala
@@ -11,7 +11,6 @@ import akka.grpc.gen.javadsl.JavaBothCodeGenerator
 import akka.grpc.gen.javadsl.JavaClientCodeGenerator
 import akka.grpc.gen.javadsl.JavaServerCodeGenerator
 import akka.grpc.gen.javadsl.play.{ PlayJavaClientCodeGenerator, PlayJavaServerCodeGenerator }
-import akka.grpc.gen.scaladsl.ScalaBothCodeGenerator
 import akka.grpc.gen.scaladsl.ScalaClientCodeGenerator
 import akka.grpc.gen.scaladsl.ScalaServerCodeGenerator
 import akka.grpc.gen.scaladsl.play.{ PlayScalaClientCodeGenerator, PlayScalaServerCodeGenerator }
