--- codegen/src/main/scala/akka/grpc/gen/scaladsl/ScalaBothCodeGenerator.scala
@@ -1,9 +1,0 @@
-/*
- * Copyright (C) 2018 Lightbend Inc. <https://www.lightbend.com>
- */
-
-package akka.grpc.gen.scaladsl
-
-object ScalaBothCodeGenerator extends ScalaServerCodeGenerator(powerApis = false) with ScalaClientCodeGenerator {
-  override def name = "akka-grpc-scaladsl-both"
-}
