+++ codegen/src/main/twirl/templates/PlayScala/AkkaGrpcClientModule.scala.txt
@@ -5,7 +5,7 @@
 @(packageName: String, services: Seq[akka.grpc.gen.scaladsl.Service])
 
 @akka.grpc.gen.Constants.DoNotEditComment
-package @{packageName}
+@{if (packageName.nonEmpty) s"package $packageName" else ""}
 
 import play.api.inject.Binding
 import play.api.{Configuration, Environment}
