+++ sbt-plugin/src/main/scala/akka/grpc/sbt/AkkaGrpcPlugin.scala
@@ -8,8 +8,9 @@ import java.io.{ByteArrayOutputStream, PrintStream}
 
 import akka.grpc.gen.CodeGenerator.ScalaBinaryVersion
 import akka.grpc.gen.scaladsl.play.{PlayScalaClientCodeGenerator, PlayScalaServerCodeGenerator}
-import akka.grpc.gen.scaladsl.{ScalaClientCodeGenerator, ScalaServerCodeGenerator}
+import akka.grpc.gen.scaladsl.{ScalaClientCodeGenerator, ScalaServerCodeGenerator, ScalaTraitCodeGenerator}
 import akka.grpc.gen.{Logger => GenLogger}
+import protocbridge.Generator
 import sbt.Keys._
 import sbt.{GlobFilter, _}
 import sbtprotoc.ProtocPlugin
@@ -153,18 +154,22 @@ object AkkaGrpcPlugin extends AutoPlugin {
     // we have a default flat_package, but that doesn't play with the java generator (it fails)
     def JavaGenerator: protocbridge.Generator = PB.gens.java
 
-    val serverPowerApis = options.contains(GeneratorOption.ServerPowerApis.setting)
+    lazy val serverPowerApis = options.contains(GeneratorOption.ServerPowerApis.setting)
+    lazy val baseGenerators: Seq[Generator] = Seq(ScalaGenerator, toGenerator(ScalaTraitCodeGenerator, scalaBinaryVersion, logger))
 
-    (for {
+    val generators = (for {
       stub <- stubs
       language <- languages
     } yield (stub, language) match {
       case (_, Java) => logger.error("Livongo: java code generators not yet available"); Seq.empty
-      case (Client, Scala) => Seq(ScalaGenerator, toGenerator(ScalaClientCodeGenerator, scalaBinaryVersion, logger))
-      case (PlayClient, Scala) => Seq(ScalaGenerator, toGenerator(PlayScalaClientCodeGenerator, scalaBinaryVersion, logger))
-      case (Server, Scala) => Seq(ScalaGenerator, toGenerator(ScalaServerCodeGenerator(serverPowerApis), scalaBinaryVersion, logger))
-      case (PlayServer, Scala) => Seq(ScalaGenerator, toGenerator(PlayScalaServerCodeGenerator(serverPowerApis), scalaBinaryVersion, logger))
+      case (Client, Scala) => Seq(toGenerator(ScalaClientCodeGenerator, scalaBinaryVersion, logger))
+      case (PlayClient, Scala) => Seq(toGenerator(PlayScalaClientCodeGenerator, scalaBinaryVersion, logger))
+      case (Server, Scala) => Seq(toGenerator(ScalaServerCodeGenerator(serverPowerApis), scalaBinaryVersion, logger))
+      case (PlayServer, Scala) => Seq(toGenerator(PlayScalaServerCodeGenerator(serverPowerApis), scalaBinaryVersion, logger))
     }).flatten.distinct
+
+    if (generators.nonEmpty) baseGenerators ++ generators
+    else generators
   }
 
   // this transforms the Akka gRPC API generators to the right protocbridge type
